requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
filelock==3.18.0
fsspec==2025.3.2
idna==3.10
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.2.4
openai-whisper==20240930
pycparser==2.22
regex==2024.11.6
requests==2.32.3
scipy==1.15.2
setuptools==75.8.0
sounddevice==0.5.1
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.2
urllib3==2.4.0
wheel==0.45.1
